pyproject.toml
[project]
name = "aipod"
version = "0.1.0"
description = "AI Podcast Agent - Web Scraper and Script Generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[rich]>=0.12.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.11",
    "trafilatura>=1.8.0",
    "psycopg[binary]>=3.1.0",
    "psycopg_pool>=3.2.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "rich>=13.7.0",
    "pendulum>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
aipod = "aipod.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["aipod"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
disallow_untyped_defs = true